Implemented button navigation on killscreen
MenuController.cs is not needed and the entire thing could be done with built in thing. To fix the issue of the kill screen not being navigatable is you had to select the first button when you pulled the menu up, otherwise the selection was still on the pause menu by default, since both are within the same scene.