Blue LVL Background
Order and how to put together the background pieces for blue level.
Scene_Scroll_BlueLVLV1_1 to V4_ 1 have to be in order by number same with the front ones.
BUT FrontV1_1 can be placed in front of Scroll_BlueLVLV3_1 or 2_1 or 4_1 or 1_1, doesnt matter as long as they are in order.